Introduce process envelope boundary for suspend protocol isolation
- Add process_envelope.hpp/cpp owning all raw SuspendRecord protocol
- Add trigger_record.hpp extracting trigger vocabulary from suspend_record.hpp
- Add process_requests.hpp with semantic DelayRequest/WaitRequest/etc types
- Move all LyraSuspend* C ABI helpers from simulation.cpp to envelope
- Replace HandleSuspendRecord with envelope-internal request decoding
- Engine exposes HandleWaitRequest primitive, not protocol-shaped consumer
- DispatchAndHandleActivation is the single public envelope entrypoint
- Decouple HasPostActivationReconciliation from suspend_records_ storage
